Crate amethyst_locale

amethyst_locale

Localisation binding a Fluent file to an Asset via the use of amethyst_assets.

Modules

resolve
The ResolveValue trait resolves Fluent AST nodes to FluentValues.

types
The FluentValue enum represents values which can be formatted to a String.

Macros

fluent_args
A helper macro to simplify creation of FluentArgs.

Structs

FluentBundle
A collection of localization messages for a single locale, which are meant to be used together in a single view, widget or any other UI abstraction.

FluentMessage
A single localization unit composed of an identifier, value, and attributes.

FluentResource
A resource containing a list of localization messages.

Locale
A loaded locale.

LocaleFormat
Loads the strings from localisation files.

Enums

FluentError

FluentValue

Type Definitions

FluentArgs
A map of arguments passed from the code to the localization to be used for message formatting.

LocaleHandle
A handle to a locale.